requirements.txt
pandas
seaborn
tda-api
flask
streamlit
